refactor(executor,helps): move claude replay helpers to helps
