+++ internal/runtime/executor/helps/claude_thinking_replay_session.go
@@ -10,6 +10,7 @@ import (
 	"net/http"
 	"strings"
 
+	claudeauth "github.com/router-for-me/CLIProxyAPI/v7/internal/auth/claude"
 	cliproxyauth "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/auth"
 	cliproxyexecutor "github.com/router-for-me/CLIProxyAPI/v7/sdk/cliproxy/executor"
 	"github.com/tidwall/gjson"
@@ -66,7 +67,7 @@ func ClaudeThinkingReplayConversationSessionKey(auth *cliproxyauth.Auth, req cli
 	if auth != nil {
 		if id := strings.TrimSpace(auth.ID); id != "" {
 			hashString(h, id)
-		} else if apiKey, _ := claudeCredentialKey(auth); apiKey != "" {
+		} else if apiKey, _ := ClaudeCredentialKey(auth); apiKey != "" {
 			hashString(h, apiKey)
 		} else {
 			hashString(h, "")
@@ -75,10 +76,10 @@ func ClaudeThinkingReplayConversationSessionKey(auth *cliproxyauth.Auth, req cli
 		hashString(h, "")
 	}
 
-	hashString(h, metadataString(opts.Metadata, cliproxyexecutor.CallerScopeMetadataKey))
-	hashString(h, metadataString(req.Metadata, cliproxyexecutor.CallerScopeMetadataKey))
-	hashString(h, metadataString(opts.Metadata, cliproxyexecutor.DerivedSessionIDMetadataKey))
-	hashString(h, metadataString(req.Metadata, cliproxyexecutor.DerivedSessionIDMetadataKey))
+	hashString(h, MetadataString(opts.Metadata, cliproxyexecutor.CallerScopeMetadataKey))
+	hashString(h, MetadataString(req.Metadata, cliproxyexecutor.CallerScopeMetadataKey))
+	hashString(h, MetadataString(opts.Metadata, cliproxyexecutor.DerivedSessionIDMetadataKey))
+	hashString(h, MetadataString(req.Metadata, cliproxyexecutor.DerivedSessionIDMetadataKey))
 
 	// Read identity headers case-insensitively so callers that supply lowercase
 	// keys (e.g. x-codex-client-id) are not collapsed with missing values.
@@ -142,15 +143,18 @@ func hashBytes(h hash.Hash, b []byte) {
 	h.Write(b)
 }
 
-// claudeCredentialKey returns the most identifying credential value available
+// ClaudeCredentialKey returns the most identifying credential value available
 // for an auth without importing the executor package.
-func claudeCredentialKey(auth *cliproxyauth.Auth) (apiKey, baseURL string) {
+func ClaudeCredentialKey(auth *cliproxyauth.Auth) (apiKey, baseURL string) {
 	if auth == nil {
 		return "", ""
 	}
 	if auth.Attributes != nil {
 		apiKey = auth.Attributes["api_key"]
 		baseURL = auth.Attributes["base_url"]
+	}
+	if apiKey == "" {
+		apiKey = claudeauth.ReadMetadataString(&auth.Metadata, "access_token")
 	}
 	return apiKey, baseURL
 }
