+++ internal/runtime/executor/helps/derived_session.go
@@ -29,7 +29,7 @@ func DerivedSessionUUID(provider string, metadataSets ...map[string]any) string 
 // ProviderSessionUUID prefers a long-lived execution session and falls back to the derived identity.
 func ProviderSessionUUID(provider string, metadataSets ...map[string]any) string {
 	for _, metadata := range metadataSets {
-		if executionID := metadataString(metadata, cliproxyexecutor.ExecutionSessionMetadataKey); executionID != "" {
+		if executionID := MetadataString(metadata, cliproxyexecutor.ExecutionSessionMetadataKey); executionID != "" {
 			return stableProviderSessionUUID(provider, "execution-session", executionID)
 		}
 	}
@@ -57,7 +57,7 @@ func DerivedAntigravitySessionID(metadataSets ...map[string]any) string {
 	return "-" + strconv.FormatInt(value, 10)
 }
 
-func metadataString(metadata map[string]any, key string) string {
+func MetadataString(metadata map[string]any, key string) string {
 	if metadata == nil {
 		return ""
 	}
